actually check we can run scm command for roles
a better error message than file not found
